+++ quantforge/__init__.py
@@ -193,6 +193,7 @@ from .holee import (
     holee_zero_coupon_bond, holee_zero_coupon_yield, holee_bond_greeks,
     holee_bond_option, holee_coupon_bond_option, holee_swaption,
     holee_caplet, holee_floorlet, holee_cap, holee_floor,
+    holee_expected_rate, holee_rate_variance,
 )
 from .overhedge import (
     Overhedge, digital_call_overhedge, digital_put_overhedge, overhedge_payoff,
@@ -338,7 +339,7 @@ try:  # pragma: no cover - trivial availability branch
 except ImportError:  # pragma: no cover
     HAS_NUMPY = False
 
-__version__ = "1.326.0"
+__version__ = "1.327.0"
 
 __all__ = [
     "OptionType",
@@ -638,6 +639,8 @@ __all__ = [
     "holee_floorlet",
     "holee_cap",
     "holee_floor",
+    "holee_expected_rate",
+    "holee_rate_variance",
     "Overhedge",
     "digital_call_overhedge",
     "digital_put_overhedge",